Extract simd code
