+++ wasm/masonry/src/lib.rs
@@ -10,3 +10,4 @@ mod layout;
 mod masonry_worker;
 mod sync;
 mod util;
+mod wide;
